[RED] remove page and map to create tests for IndonesiaMap component
